+++ Sources/KeyboardKit/_Deprecated/DeviceSpecificInputSetProvider+Deprecated.swift
@@ -2,25 +2,25 @@ import Foundation
 
 public extension DeviceSpecificInputSetProvider {
     
-    @available(*, deprecated, message: "Use KeyboardInputRow initializer instead")
-    func row(_ chars: String) -> KeyboardInputRow {
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
+    func row(_ chars: String) -> InputSetRow {
         row(chars.chars)
     }
     
-    @available(*, deprecated, message: "Use KeyboardInputRow initializer instead")
-    func row(_ chars: [String]) -> KeyboardInputRow {
-        KeyboardInputRow(chars)
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
+    func row(_ chars: [String]) -> InputSetRow {
+        InputSetRow(chars)
     }
     
-    @available(*, deprecated, message: "Use KeyboardInputRow initializer instead")
-    func row(lowercased: [String], uppercased: [String]) -> KeyboardInputRow {
-        KeyboardInputRow(
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
+    func row(lowercased: [String], uppercased: [String]) -> InputSetRow {
+        InputSetRow(
             lowercased: lowercased,
             uppercased: uppercased)
     }
     
-    @available(*, deprecated, message: "Use KeyboardInputRow initializer instead")
-    func row(lowercased: String, uppercased: String) -> KeyboardInputRow {
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
+    func row(lowercased: String, uppercased: String) -> InputSetRow {
         row(lowercased: lowercased.chars,
             uppercased: uppercased.chars)
     }