+++ Sources/KeyboardKit/_Deprecated/KeyboardInput.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "InputSetItem")
+public typealias KeyboardInput = InputSetItem